replaced the success image with an icon, renamed success file and logo
