+++ src/Global Components/Modals/success-modal/Success.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
-import successImg from '../../assets/Images/teenyicons_tick-circle-solid.png'
 import { useNavigate } from 'react-router-dom'
+import { FaCheckCircle } from "react-icons/fa";
 
 type successProps = {
     mode: string
@@ -10,7 +10,7 @@ const Success = ({mode}: successProps) => {
     return (
         <div className='add-success'>
             <h2>Quiz Successfully {(mode === 'add') && "Added"} {(mode === 'edit') && "Edited"}</h2>
-            <img src={successImg} alt="" />
+            <FaCheckCircle color={'#1FAF38'} size={150} style={{marginBottom: '20px'}}/>
             <p>You have successfully {(mode === 'add') && "added"} {(mode === 'edit') && "edited"} that quiz</p>
             <button className="finish" onClick={() => navigate('/select-quiz')}>FINISH</button>
         </div>
