+++ src/Pages/Selectquiz/Index.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import logo from '../../assets/Images/Group 1171275826.png'
+import logo from '../../assets/Images/logo.png'
 import "./styles.css"
 import { LuPlusCircle } from "react-icons/lu";
 import { GoPencil } from "react-icons/go";
