+++ src/Pages/Subadmin/account_settings/AccountSettings.tsx
@@ -1,6 +1,5 @@
 import React from 'react'
 import './accountSettings.css'
-import Sidebar from './sidebar/Sidebar'
 import { Outlet } from 'react-router-dom'
 import { FiUser } from "react-icons/fi";
 import { RiLockPasswordLine } from "react-icons/ri";
@@ -15,11 +14,10 @@ const AccountSettings = () => {
 
   return (
     <div className='account-setting-container'>
-        <Sidebar />
         <Outlet />
         <div className="right-section">
-            <button className={(location.pathname === '/account-settings/profile-settings' || location.pathname === '/account-settings') ? 'active' : ''}  onClick={() => naviagte('/account-settings/profile-settings')}><FiUser size={20}/> Profile Settings <IoIosArrowForward size={20}/></button>
-            <button className={(location.pathname === '/account-settings/password-settings') ? 'active' : ''} onClick={() => naviagte('/account-settings/password-settings')}><RiLockPasswordLine size={20} /> Password Settings <IoIosArrowForward size={20}/></button>
+            <button className={(location.pathname === '/subadmin/account-settings/profile-settings' || location.pathname === '/subadmin/account-settings') ? 'active' : ''}  onClick={() => naviagte('/subadmin/account-settings/profile-settings')}><FiUser size={20}/> Profile Settings <IoIosArrowForward size={20}/></button>
+            <button className={(location.pathname === '/subadmin/account-settings/password-settings') ? 'active' : ''} onClick={() => naviagte('/subadmin/account-settings/password-settings')}><RiLockPasswordLine size={20} /> Password Settings <IoIosArrowForward size={20}/></button>
             <button><MdLogout size={20} /> logout <IoIosArrowForward size={20}/></button>
         </div>
     </div>
