+++ src/Pages/Subadmin/account_settings/accountSettings.css
@@ -1,8 +1,10 @@
 .account-setting-container{
     display: flex;
+    height: 100%;
+    align-items: center;
 }
 .right-section{
-    width: calc(54% - 255px);
+    width: 40%;
     display: flex;
     flex-direction: column;
     justify-content: space-between;
