+++ src/Pages/Subadmin/account_settings/password_settings/passwordSettings.css
@@ -1,5 +1,5 @@
 .password-settings{
-    width: 48%;
+    width: 60%;
     margin: 90px 0;
     border-right: 1px solid rgba(51, 51, 51, 0.3);
     padding:  50px;
