--- src/Pages/Subadmin/account_settings/sidebar/sidebar.css
@@ -1,93 +1,0 @@
-.admin-sidebar{
-    width: 255px;
-    min-height: 100vh;
-    background-color: #EDEDED;
-    padding: 50px 0;
-}
-.admin-sidebar .head{
-    display: flex;
-    width: 100%;
-    align-items: center;
-    gap: 13px;
-    padding-left: 40px;
-    margin-top: 30px;
-    font-size: 20px;
-}
-.admin-sidebar .head:first-child{
-    margin-top: 0;
-    margin-bottom: 10px;
-}
-.admin-sidebar .head img{
-    width: 24px;
-}
-.admin-sidebar .head span{
-    font-size: 20px;
-    font-weight: 500;
-    color: #333333;
-}
-.admin-sidebar .sidebar-items{
-    margin-top: 20px;
-}
-.admin-sidebar .sidebar-items li{
-    width: 100%;
-    list-style-type: none;
-    background-color: #fff;
-    padding: 10px 0 10px 40px;
-    margin-bottom: 20px;
-    position: relative;
-    transition: .1s;
-    display: flex;
-    align-items: center;
-    gap: 15px;
-    cursor: default;
-}
-.admin-sidebar .sidebar-items li:hover, .admin-sidebar .sidebar-items li.active{
-    box-shadow: 2px 2px 10px rgba(0, 0, 0, 0.2);
-}
-.admin-sidebar .sidebar-items li::before{
-    content: '';
-    position: absolute;
-    left: 0;
-    height: 100%;
-    width: 5px;
-    background: transparent;    
-    top: 0;
-    transition: .1s;
-}
-.admin-sidebar .sidebar-items li:hover::before, .admin-sidebar .sidebar-items li.active::before{
-    background-color: #8F19E7;
-}
-
-.admin-sidebar .sidebar-items li span{
-    color: #333333;
-    opacity: 0.8;
-    font-size: 16px;
-    font-weight: 400;
-}
-.admin-sidebar .sidebar-items li.active svg, .admin-sidebar .sidebar-items li.active span{
-    color: #8F19E7;
-    opacity: 1;
-    font-weight: 400;
-}
-.admin-sidebar .edit-add{
-    margin-left: 40px;
-    padding: 6px;
-    border-radius: 7px;
-    background-color: #8F19E7;
-    display: flex;
-    width: fit-content;
-    align-items: center;
-    color: #fff;
-}
-.admin-sidebar .edit-add .edit-add-icons{
-    margin-left: 8px;
-    margin-right: 5px;
-    width: 20px;
-    height: 20px;
-    background-color: #fff;
-    color: #000000;
-    border-radius: 5px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
